refactor: 인터셉터 추상화처리 건 기존 로직과 대체 처리
