+++ src/main/java/nextstep/auth/application/UserDetailService.java
@@ -1,0 +1,7 @@
+package nextstep.member.application;
+
+import nextstep.member.domain.LoginMember;
+
+public interface UserDetailService {
+	LoginMember loadUserByUsername(String email);
+}
