--- src/main/java/nextstep/auth/authentication/SessionAuthenticationInterceptor_new.java
@@ -1,30 +1,0 @@
-package nextstep.auth.authentication;
-
-import nextstep.auth.context.Authentication;
-import nextstep.auth.context.SecurityContext;
-import nextstep.member.application.CustomUserDetailsService;
-import nextstep.member.domain.LoginMember;
-import org.springframework.web.servlet.HandlerInterceptor;
-
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
-import javax.servlet.http.HttpSession;
-import java.util.Map;
-
-import static nextstep.auth.context.SecurityContextHolder.SPRING_SECURITY_CONTEXT_KEY;
-
-public class SessionAuthenticationInterceptor_new extends AuthenticationInterceptor {
-    public static final String USERNAME_FIELD = "username";
-    public static final String PASSWORD_FIELD = "password";
-
-    public SessionAuthenticationInterceptor_new(CustomUserDetailsService userDetailsService, AuthenticationConverter converter) {
-        super(userDetailsService, converter);
-    }
-
-    @Override
-    public void afterAuthentication(HttpServletRequest request, HttpServletResponse response, Authentication authentication) {
-        HttpSession httpSession = request.getSession();
-        httpSession.setAttribute(SPRING_SECURITY_CONTEXT_KEY, new SecurityContext(authentication));
-        response.setStatus(HttpServletResponse.SC_OK);
-    }
-}
