--- src/main/java/nextstep/auth/authentication/TokenAuthenticationInterceptor_new.java
@@ -1,49 +1,0 @@
-package nextstep.auth.authentication;
-
-import com.fasterxml.jackson.core.JsonProcessingException;
-import com.fasterxml.jackson.databind.ObjectMapper;
-import nextstep.auth.context.Authentication;
-import nextstep.auth.token.JwtTokenProvider;
-import nextstep.auth.token.TokenRequest;
-import nextstep.auth.token.TokenResponse;
-import nextstep.member.application.CustomUserDetailsService;
-import nextstep.member.domain.LoginMember;
-import org.springframework.http.MediaType;
-import org.springframework.web.servlet.HandlerInterceptor;
-
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
-import java.io.IOException;
-
-public class TokenAuthenticationInterceptor_new extends AuthenticationInterceptor {
-
-    private JwtTokenProvider jwtTokenProvider;
-    private ObjectMapper objectMapper;
-
-    public TokenAuthenticationInterceptor_new(CustomUserDetailsService userDetailsService, AuthenticationConverter converter, JwtTokenProvider jwtTokenProvider, ObjectMapper objectMapper) {
-        super(userDetailsService, converter);
-        this.jwtTokenProvider = jwtTokenProvider;
-        this.objectMapper = objectMapper;
-    }
-
-    @Override
-    public void afterAuthentication(HttpServletRequest request, HttpServletResponse response, Authentication authentication) throws IOException {
-        TokenResponse tokenResponse = generateToken(authentication);
-
-        setToken(response, tokenResponse);
-    }
-
-    private TokenResponse generateToken(Authentication authentication) throws JsonProcessingException {
-        String payload = objectMapper.writeValueAsString(authentication.getPrincipal());
-        String token = jwtTokenProvider.createToken(payload);
-
-        return TokenResponse.of(token);
-    }
-
-    private void setToken(HttpServletResponse response, TokenResponse tokenResponse) throws IOException {
-        String responseToClient = objectMapper.writeValueAsString(tokenResponse);
-        response.setStatus(HttpServletResponse.SC_OK);
-        response.setContentType(MediaType.APPLICATION_JSON_VALUE);
-        response.getOutputStream().print(responseToClient);
-    }
-}
