+++ src/main/java/nextstep/auth/domain/UserDetail.java
@@ -1,0 +1,2 @@
+package nextstep.auth;public class UserDetail {
+}
