--- src/test/java/nextstep/subway/unit/TokenAuthenticationInterceptor_newTest.java
@@ -1,87 +1,0 @@
-package nextstep.subway.unit;
-
-import com.fasterxml.jackson.databind.ObjectMapper;
-import nextstep.auth.authentication.*;
-import nextstep.auth.context.Authentication;
-import nextstep.auth.token.JwtTokenProvider;
-import nextstep.auth.token.TokenRequest;
-import nextstep.auth.token.TokenResponse;
-import nextstep.member.application.CustomUserDetailsService;
-import nextstep.member.domain.LoginMember;
-import org.junit.jupiter.api.Test;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.MediaType;
-import org.springframework.mock.web.MockHttpServletRequest;
-import org.springframework.mock.web.MockHttpServletResponse;
-
-import java.io.IOException;
-
-import static org.assertj.core.api.Assertions.assertThat;
-import static org.mockito.ArgumentMatchers.anyString;
-import static org.mockito.Mockito.*;
-
-class TokenAuthenticationInterceptor_newTest {
-    private static final String EMAIL = "email@email.com";
-    private static final String PASSWORD = "password";
-    public static final String JWT_TOKEN = "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIiLCJuYW1lIjoiSm9obiBEb2UiLCJpYXQiOjE1MTYyMzkwMjJ9.ih1aovtQShabQ7l0cINw4k1fagApg3qLWiB8Kt59Lno";
-
-    private final CustomUserDetailsService userDetailsService = mock(CustomUserDetailsService.class);
-    private final TokenAuthenticationConverter converter = new TokenAuthenticationConverter();
-    private final JwtTokenProvider jwtTokenProvider = mock(JwtTokenProvider.class);
-    private final ObjectMapper objectMapper = spy(ObjectMapper.class);
-
-    @Test
-    void convert() throws IOException {
-        // given
-        TokenAuthenticationInterceptor_new interceptor = new TokenAuthenticationInterceptor_new(userDetailsService, converter, jwtTokenProvider, objectMapper);
-        MockHttpServletRequest request = createMockRequest();
-
-        // when
-        AuthenticationToken authenticationToken = interceptor.convert(request);
-
-        // then
-        assertThat(authenticationToken.getPrincipal()).isEqualTo(EMAIL);
-        assertThat(authenticationToken.getCredentials()).isEqualTo(PASSWORD);
-    }
-
-    @Test
-    void authenticate() {
-        // given
-        TokenAuthenticationInterceptor_new interceptor = new TokenAuthenticationInterceptor_new(userDetailsService, converter, jwtTokenProvider, objectMapper);
-        when(userDetailsService.loadUserByUsername(EMAIL)).thenReturn(new LoginMember(1L, EMAIL, PASSWORD, 20));
-
-        // when
-        AuthenticationToken authenticationToken = AuthenticationToken.of(EMAIL, PASSWORD);
-        Authentication authentication = interceptor.authenticate(authenticationToken);
-
-        // then
-        assertThat(authentication.getPrincipal()).isNotNull();
-    }
-
-    @Test
-    void preHandle() throws IOException {
-        // given
-        TokenAuthenticationInterceptor_new interceptor = new TokenAuthenticationInterceptor_new(userDetailsService, converter, jwtTokenProvider, objectMapper);
-
-        when(userDetailsService.loadUserByUsername(EMAIL)).thenReturn(new LoginMember(1L, EMAIL, PASSWORD, 20));
-        when(jwtTokenProvider.createToken(anyString())).thenReturn(JWT_TOKEN);
-
-        // when
-        MockHttpServletRequest request = createMockRequest();
-        MockHttpServletResponse response = new MockHttpServletResponse();
-        interceptor.preHandle(request, response, new Object());
-
-        // then
-        assertThat(response.getStatus()).isEqualTo(HttpStatus.OK.value());
-        assertThat(response.getContentType()).isEqualTo(MediaType.APPLICATION_JSON_VALUE);
-        assertThat(response.getContentAsString()).isEqualTo(new ObjectMapper().writeValueAsString(TokenResponse.of(JWT_TOKEN)));
-    }
-
-    private MockHttpServletRequest createMockRequest() throws IOException {
-        MockHttpServletRequest request = new MockHttpServletRequest();
-        TokenRequest tokenRequest = new TokenRequest(EMAIL, PASSWORD);
-        request.setContent(new ObjectMapper().writeValueAsString(tokenRequest).getBytes());
-        return request;
-    }
-
-}
